requirements.txt
uvicorn==0.30.6
fastapi==0.115.0
python-multipart==0.0.12
opencv-python-headless==4.10.0.84
numpy==1.26.4
Pillow==10.4.0
mysql-connector-python>=8.0.0
torch>=1.7.0
torchvision>=0.8.1
pandas>=1.1.4
requests>=2.23.0
scipy>=1.4.1
tqdm>=4.64.0
seaborn>=0.11.0
